relates to #31, #33, #46, #48 Data models are made and working with game logic
This Commit handles multiple issues form 2 different user-stories, since they were very much intertwined and could hardly be handled separately. Now, the Models for the Player-Stats are created and are in working order. I made Changes to the Card-Class, since a few attributes were not needed (already handled in Effect Class). Also I refactored TurnManager so that the method calls could be more concise and readable.