implement skeleton structure for client-server design
relates to #95: all data for the UI has to be communicated over the network anyways,
so the present classes have to provide according interfaces for the ClientConnector
to call the necessary methods in response to a message from the server. Same thing
goes for the backend: all mesages received will be mapped to actions on the whole
gamestate in the MessageHandler

relates to #100, #94: message logic can be handled by the client in the ClientConnector,
by the server in the MessageHandler

relates to #99: MessageTypes will be defined in the corresponding class, the Message
class might handle conversion to and from a transmission format (e.g. json with Gson)

relates to #98: the server thread is contained in the ServerThread class