Implemented first prototype of Variable Card Design.
The goal is to build a view of a card programmatically, in order to display all stats of the card. This CardView should be scalable to support all new types of Cards.

Currently not working, since the edited Scaling of the View did not affect the size that android perceived. An attempt to fix this by Overriding OnMeasure() has left 3/4th of the Card not visible.