added id to Card and playCardById to PlayArea
It was necessary to give IDs to the cards, since cards will be referenced by them in the Client-Server-Communication. Hence it was also necessary to implement a corresponding method in the playArea, that would take said Card ID, map it to the desired Card-Object and then call the corresponding method with the newly found card-Object as parameter.