Preparations for Server to handle GameSession
Since we need an instance that will keep track and handle/control the GameSession, I decided
to reference it in the ServerThread. Since other classes must have access to the GameSession,
the ServerThread makes use of a semi “Singleton” Pattern and offers a way to get access to it
by letting other classes access it via the static method getInstace, that returns the only
instance of the ServerThread.
Also I started to implement the Methods to handle the incoming Messages on the server side.