Keep track of all cards in Card-Class
Since we need to share with all clients, which cards we have in the game
and what IDs, effects etc. they have, I decided to have a static Deck of Cards
in the Card-Class, to which every Card created is added to via the constructor.
Other classes can access this collection for distribution of Deck information
Via the network later on.