Changes to ServerThread to communicate initial game state
The server now creates new Players and set their Cards as well as the
Market cards and then creates a new GameSession. After this, it will
Send to all clients, the full set of Cards in the game as well as initial
Player stats and who has the first turn.