implementation of Champion-class
Up until now, champions were not part of the game. Since this will
change in sprint 3, I now implemented the required attributes and
methods the class needs in order to function correctly in accordance with
the game rules. Champions extend Cards, but they now have their own
health and can be killed. The also can be expended in the players turn,
which would reapply their effects (handled by PlayArea).