accommodate for new Champion class
Up until now, champions were not part of the game. Since this will
change in sprint 3, the PlayArea needs to implement some new functionality
to account for the champions. Champions are not played to the playedCards
but rather to the playedChampions and are not cleared at the end of a turn
like the other cards. Also they can be expended and attacked by the opponent.