[#119]changes to tests
Since many changes were made to the backend of the game, the refactorings also entailed
changes in the test classes. Most of them were not changed otherwise except the champion
test. But also many tests are currently not working, this will be addressed in the test branch
for this feature that is to come.