[#119] Added Champions to PlayArea
The champions occupy an extra area on the play field of the game and stay
there when played after the end of the turn. Therefore,
they need to be stored and handled separately from normal cards.
I added the needed functionality to handle the attacking, expanding, playing and
resetting of the champions. Also, I made changes to the playCard() to now better
suit the addition of the champions and the way synergies are triggered with them
in mind.