[#119] added PlayedChampion-View
This view works as the UI element as required by the
MVP model approach we took for our game. It comes with the necessary methods
to act upon and make necessary changes to the UI for all actions that the champion
cards could need (i.e. update (includes add, remove by checking the full list of cards
to display stored by its presenter), expend, reset).