[#119] added PlayedChampion-Presenter
This Presenter works as the linking element between or server-related Backend
(handled in the mainAct.) and the PlayedChampion-View as required by the
MVP model approach we took for our game. It comes with the necessary methods
to act upon and make necessary changes to the UI for all actions that the champion
cards could need (i.e. add, remove, expend, reset).