[#10]small adjustments to music player
I put the code that starts the music before the code that
initializes the views, so that if the players must wait,
they can at least enjoy it. Also, since the song positon was
saved in the shared-preferences, it would not start at the beginning, even when the game fully restarts. This changed so
now it will only continue, when the app was minimized.